Renamed packages com.jblosc to org.blosc
