--- jblosc/src/main/java/com/jblosc/Shuffle.java
@@ -1,7 +1,0 @@
-package com.jblosc;
-
-public class Shuffle {
-	public static int NO_SHUFFLE = 0;
-	public static int BYTE_SHUFFLE = 1;
-	public static int BIT_SHUFFLE = 2;
-}
