+++ jblosc/src/main/java/org/blosc/Shuffle.java
@@ -1,0 +1,7 @@
+package org.blosc;
+
+public class Shuffle {
+	public static int NO_SHUFFLE = 0;
+	public static int BYTE_SHUFFLE = 1;
+	public static int BIT_SHUFFLE = 2;
+}
